requirements.txt
pandas>=1.5.0
faker>=19.0.0
google-cloud-bigquery>=3.10.0
google-cloud-storage>=2.10.0
matplotlib>=3.7.0
